refactor: #332 SearchUseCase 프로토콜 분리
